feat(domain): ajustar relaciones, tipos decimal y valores por defecto en entidades

+++ Meseta_Verda.Domain/Entities/DetallePedido.cs
@@ -19,6 +19,6 @@ namespace Meseta_Verda.Domain.Entities
         
         public float Cantidad { get; set; }
         public float PrecioUnitario { get; set; }
-        public float Subtotal { get; set; }
+        public decimal Subtotal { get; set; }
     }
 }
